# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bake Delight Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  imageUrl?: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface CreateCategoryBody {
  name: string;
  slug: string;
  description?: string | null;
  imageUrl?: string | null;
  sortOrder?: number;
}

export type ProductVariantOptionsItem = {
  label: string;
  priceAdjustment: number;
};

export interface ProductVariant {
  name: string;
  /** e.g. size, flavor, frosting */
  type: string;
  options: ProductVariantOptionsItem[];
}

export interface ProductAddon {
  name: string;
  price: number;
}

export interface Product {
  id: number;
  categoryId?: number | null;
  category?: Category | null;
  name: string;
  slug: string;
  description?: string | null;
  basePrice: number;
  imageUrls: string[];
  variants: ProductVariant[];
  addons: ProductAddon[];
  allowCustomMessage: boolean;
  isVisible: boolean;
  isAvailable: boolean;
  orderCount: number;
  createdAt: string;
}

export interface CreateProductBody {
  categoryId?: number | null;
  name: string;
  slug: string;
  description?: string | null;
  basePrice: number;
  imageUrls?: string[];
  variants?: ProductVariant[];
  addons?: ProductAddon[];
  allowCustomMessage?: boolean;
  isVisible?: boolean;
  isAvailable?: boolean;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  in_baking: "in_baking",
  out_for_delivery: "out_for_delivery",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type OrderItemSelectedVariants = { [key: string]: string };

export interface OrderItem {
  productId: number;
  productName: string;
  productImageUrl?: string | null;
  quantity: number;
  unitPrice: number;
  selectedVariants: OrderItemSelectedVariants;
  selectedAddons: string[];
  customMessage?: string | null;
  subtotal: number;
}

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  customerEmail?: string | null;
  deliveryAddress: string;
  deliveryDate: string;
  deliveryTimeSlot?: string | null;
  notes?: string | null;
  status: OrderStatus;
  items: OrderItem[];
  subtotal: number;
  discountAmount: number;
  totalAmount: number;
  couponCode?: string | null;
  whatsappMessageSent: boolean;
  createdAt: string;
}

export type CreateOrderBodyItemsItemSelectedVariants = {
  [key: string]: string;
};

export type CreateOrderBodyItemsItem = {
  productId: number;
  quantity: number;
  selectedVariants: CreateOrderBodyItemsItemSelectedVariants;
  selectedAddons: string[];
  customMessage?: string | null;
};

export interface CreateOrderBody {
  customerName: string;
  customerPhone: string;
  customerEmail?: string | null;
  deliveryAddress: string;
  deliveryDate: string;
  deliveryTimeSlot?: string | null;
  notes?: string | null;
  couponCode?: string | null;
  items: CreateOrderBodyItemsItem[];
}

export interface UpdateOrderStatusBody {
  status: OrderStatus;
}

export type OrdersSummaryStatusBreakdownItem = {
  status: OrderStatus;
  count: number;
};

export interface OrdersSummary {
  totalOrders: number;
  todayOrders: number;
  pendingOrders: number;
  totalRevenue: number;
  todayRevenue: number;
  statusBreakdown: OrdersSummaryStatusBreakdownItem[];
  recentOrders: Order[];
}

export interface CalendarEntry {
  date: string;
  orderCount: number;
  orders: Order[];
}

export interface CapacityStatus {
  available: boolean;
  remainingSlots: number;
  isBlackout: boolean;
}

export interface Banner {
  id: number;
  title: string;
  subtitle?: string | null;
  imageUrl: string;
  linkUrl?: string | null;
  isActive: boolean;
  sortOrder: number;
  createdAt: string;
}

export interface CreateBannerBody {
  title: string;
  subtitle?: string | null;
  imageUrl: string;
  linkUrl?: string | null;
  isActive?: boolean;
  sortOrder?: number;
}

export interface BlackoutDate {
  id: number;
  date: string;
  reason?: string | null;
  createdAt: string;
}

export interface CreateBlackoutDateBody {
  date: string;
  reason?: string | null;
}

export type CouponDiscountType =
  (typeof CouponDiscountType)[keyof typeof CouponDiscountType];

export const CouponDiscountType = {
  percentage: "percentage",
  fixed: "fixed",
} as const;

export interface Coupon {
  id: number;
  code: string;
  discountType: CouponDiscountType;
  discountValue: number;
  minOrderAmount?: number | null;
  maxUses?: number | null;
  usedCount: number;
  isActive: boolean;
  expiresAt?: string | null;
  createdAt: string;
}

export type CreateCouponBodyDiscountType =
  (typeof CreateCouponBodyDiscountType)[keyof typeof CreateCouponBodyDiscountType];

export const CreateCouponBodyDiscountType = {
  percentage: "percentage",
  fixed: "fixed",
} as const;

export interface CreateCouponBody {
  code: string;
  discountType: CreateCouponBodyDiscountType;
  discountValue: number;
  minOrderAmount?: number | null;
  maxUses?: number | null;
  isActive?: boolean;
  expiresAt?: string | null;
}

export interface ValidateCouponBody {
  code: string;
  orderAmount: number;
}

export interface CouponValidation {
  valid: boolean;
  coupon?: Coupon | null;
  discountAmount: number;
  message: string;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResult {
  success: boolean;
  message: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export type ListProductsParams = {
  categoryId?: number;
  search?: string;
  visible?: boolean;
  available?: boolean;
};

export type ListOrdersParams = {
  status?: OrderStatus;
  date?: string;
};

export type GetOrdersCalendarParams = {
  /**
   * YYYY-MM
   */
  month?: string;
};

export type CheckDeliveryCapacityParams = {
  date: string;
  timeSlot?: string;
};