# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bake Delight Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted

  # ── Products ──────────────────────────────────────────────────────────────
...
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      summary: List active banners
      responses:
        "200":
          description: List of banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create a banner (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerBody"
      responses:
        "201":
          description: Created banner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/{id}:
    put:
      operationId: updateBanner
      tags: [banners]
      summary: Update a banner (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerBody"
      responses:
        "200":
          description: Updated banner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete a banner (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted

  # ── Blackout Dates ─────────────────────────────────────────────────────────
  /blackout-dates:
    get:
      operationId: listBlackoutDates
      tags: [blackout-dates]
      summary: List blackout dates
      responses:
        "200":
          description: List of blackout dates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlackoutDate"
    post:
      operationId: createBlackoutDate
      tags: [blackout-dates]
      summary: Create a blackout date (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlackoutDateBody"
      responses:
        "201":
          description: Created blackout date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlackoutDate"
  /blackout-dates/{id}:
    delete:
      operationId: deleteBlackoutDate
      tags: [blackout-dates]
      summary: Delete a blackout date (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted

  # ── Coupons ────────────────────────────────────────────────────────────────
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List all coupons (admin)
      responses:
        "200":
          description: List of coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create a coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "201":
          description: Created coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCouponBody"
      responses:
        "200":
          description: Coupon validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidation"
  /coupons/{id}:
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete a coupon (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted

  # ── Admin Auth ─────────────────────────────────────────────────────────────
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, sortOrder, createdAt]
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required: [name, slug]

    # ── Product ───────────────────────────────────────────────────────────────
...
    Banner:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        imageUrl:
          type: string
        linkUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, imageUrl, isActive, sortOrder, createdAt]
    CreateBannerBody:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        imageUrl:
          type: string
        linkUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        sortOrder:
          type: integer
      required: [title, imageUrl]

    # ── Blackout Date ──────────────────────────────────────────────────────────
    BlackoutDate:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        reason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, date, createdAt]
    CreateBlackoutDateBody:
      type: object
      properties:
        date:
          type: string
          format: date
        reason:
          type: string
          nullable: true
      required: [date]

    # ── Coupon ─────────────────────────────────────────────────────────────────
    Coupon:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minOrderAmount:
          type: number
          nullable: true
        maxUses:
          type: integer
          nullable: true
        usedCount:
          type: integer
        isActive:
          type: boolean
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, code, discountType, discountValue, usedCount, isActive, createdAt]
    CreateCouponBody:
      type: object
      properties:
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minOrderAmount:
          type: number
          nullable: true
        maxUses:
          type: integer
          nullable: true
        isActive:
          type: boolean
        expiresAt:
          type: string
          format: date-time
          nullable: true
      required: [code, discountType, discountValue]
    ValidateCouponBody:
      type: object
      properties:
        code:
          type: string
        orderAmount:
          type: number
      required: [code, orderAmount]
    CouponValidation:
      type: object
      properties:
        valid:
          type: boolean
        coupon:
          $ref: "#/components/schemas/Coupon"
          nullable: true
        discountAmount:
          type: number
        message:
          type: string
      required: [valid, discountAmount, message]

    # ── Admin ──────────────────────────────────────────────────────────────────
    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
      required: [password]
    AdminLoginResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...